requirements.txt
chainlit>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
sounddevice>=0.4.6
numpy>=1.24.0
python-multipart>=0.0.6
google-generativeai>=0.8.0
